mark transactions as visited

+++ programs/riptide/src/account.rs
@@ -15,7 +15,7 @@ const SLOT_HASH_VEC_PREFIX_BYTES: usize = 8;
 const SLOT_HASH_ENTRY_SLOT_BYTES: usize = 8;
 const SLOT_HASH_SKIP_BYTES: usize = SLOT_HASH_VEC_PREFIX_BYTES + SLOT_HASH_ENTRY_SLOT_BYTES;
 const U64_SIZE_BYTES: usize = 8;
-const VISITED_ARR_SIZE: u8 = 30; // 30 purchases / 10 minutes => 1 purchase every 20sec.
+const VISITED_ARR_SIZE: usize = 30; // 30 purchases / 10 minutes => 1 purchase every 20sec.
 
 #[error]
 pub enum RiptideError {
@@ -128,16 +128,16 @@ pub enum CampaignState {
 pub struct VisitedQueue {
     pub next: u8,
     pub count: u8,
-    pub queue: [u32; VISITED_ARR_SIZE as usize],
+    pub queue: [u32; VISITED_ARR_SIZE],
 }
 
 impl VisitedQueue {
     fn push(&mut self, val: u32) {
         self.queue[self.next as usize] = val;
-        self.next = (self.next + 1) % VISITED_ARR_SIZE;
+        self.next = (self.next + 1) % (VISITED_ARR_SIZE as u8);
         self.count = self.count + 1;
-        if self.count > VISITED_ARR_SIZE {
-            self.count = VISITED_ARR_SIZE;
+        if self.count > VISITED_ARR_SIZE as u8 {
+            self.count = VISITED_ARR_SIZE as u8;
         }
     }
     fn has(&self, val: u32) -> bool {
@@ -146,10 +146,10 @@ impl VisitedQueue {
                 return true;
             }
         }
-        if self.count < VISITED_ARR_SIZE {
+        if self.count < VISITED_ARR_SIZE as u8 {
             return false;
         }
-        for idx in self.next..VISITED_ARR_SIZE {
+        for idx in self.next..VISITED_ARR_SIZE as u8 {
             if self.queue[idx as usize] == val {
                 return true;
             }
@@ -207,7 +207,7 @@ impl Campaign {
         self.visited = VisitedQueue {
             next: 0,
             count: 0,
-            queue: [0; VISITED_ARR_SIZE as usize],
+            queue: [0; VISITED_ARR_SIZE],
         };
         self.vaults = Vec::new();
         Ok(())
